DEV/orchestrator/security/zero_trust.py
"""
FilePath: "/DEV/orchestrator/security/zero_trust.py"
Project: Unified Bot Protocol (UBP)
Component: Security (Zero Trust, encryption, authN/Z)
Author: "Michael Landbo"
Date created: "21/12/2025"
Date Modified: "21/12/2025"
Version: "v.1.0.0"
"""

import datetime
import hashlib
import json
import logging
import os
import time
from typing import Dict, List, Optional, Tuple

import jwt
from cryptography.hazmat.primitives import hashes
from cryptography.x509 import load_pem_x509_certificate

# Logging

logger = logging.getLogger(__name__)


class ZeroTrustManager:
    """
    Initialize the ZeroTrustManager.

    self.session_store: A dictionary mapping bot IDs to their active session data.
    self.device_fingerprints: A dictionary mapping bot IDs to their device fingerprints.
    self.trust_scores: A dictionary mapping bot IDs to their trust scores.
    self.jwt_secret: A secret key used for JWT signing. Load this securely in production.
    self.certificate_store: A dictionary mapping bot IDs to their certificates.
    """

    def __init__(self):
        # In-memory stores
        self.session_store = {}
        self.device_fingerprints = {}
        self.trust_scores = {}
        # In production, load this from a secure secret manager
        self.jwt_secret = os.urandom(32)
        self.certificate_store = {}

    async def verify_identity(
        self, bot_id: str, credentials: Dict, context: Dict
    ) -> Tuple[bool, Dict]:
        """
        Implement Zero Trust verification using multiple factors.
        Returns (is_verified, context_with_results)
        """
        trust_score = 0
        verification_results = {}

        # 1. Certificate-based authentication
        cert_valid = await self._verify_certificate(
            bot_id, credentials.get("certificate")
        )
        if cert_valid:
            trust_score += 30
            verification_results["cert_auth"] = True

        # 2. Device fingerprint verification
        device_verified = await self._verify_device_fingerprint(
            bot_id, context.get("device_fingerprint")
        )
        if device_verified:
            trust_score += 20
            verification_results["device_auth"] = True

        # 3. Behavioral analysis
        behavior_score = await self._analyze_behavior(bot_id, context)
        trust_score += behavior_score
        verification_results["behavior_score"] = behavior_score

        # 4. Context validation
        context_score = await self._validate_context(context)
        trust_score += context_score
        verification_results["context_score"] = context_score

        # Store trust score
        self.trust_scores[bot_id] = trust_score

        # Threshold for verification (e.g., 70/100)
        return trust_score >= 70, {
            "trust_score": trust_score,
            "verification_results": verification_results,
        }

    async def _verify_certificate(self, bot_id: str, cert_data: Optional[str]) -> bool:
        """Verify bot's certificate"""
        if cert_data is None:
            return False

        try:
            cert = load_pem_x509_certificate(cert_data.encode("utf-8"))

            # Verify certificate is not expired
            if (
                cert.not_valid_after is None
                or cert.not_valid_after < datetime.datetime.fromtimestamp(time.time())
            ):
                return False

            # Verify the certificate matches known store
            cert_fingerprint = cert.fingerprint(hashes.SHA256()).hex()

            # In a real implementation, you would check if this fingerprint belongs to the bot_id
            # For now, we check if it exists in our store
            if cert_fingerprint in self.certificate_store[bot_id]:
                return True

        except (ValueError, TypeError) as e:
            # Handle specific exceptions
            logger.error("%(message)s", {"message": str(e)})
            return False

        return False

    async def _verify_device_fingerprint(
        self, bot_id: str, fingerprint: Optional[Dict]
    ) -> bool:
        """Verify bot's device fingerprint"""
        if not fingerprint:
            return False

        stored_fingerprint = self.device_fingerprints.get(bot_id)
        if not stored_fingerprint:
            # First time seeing this device, store it (Trust on First Use - TOFU)
            self.device_fingerprints[bot_id] = fingerprint
            return True

        # Compare fingerprint components
        match_count = sum(
            1 for k, v in fingerprint.items() if stored_fingerprint.get(k) == v
        )
        match_score = match_count / len(fingerprint) if fingerprint else 0

        return match_score >= 0.8

    async def _analyze_behavior(self, bot_id: str, context: Dict) -> int:
        """Analyze bot's behavior for anomalies"""
        bot_id = str(bot_id)
        if bot_id not in self.session_store:
            self.session_store[bot_id] = {
                "recent_commands": [],
                "command_timestamps": [],
                "resource_metrics": {},
            }

        score = 0

        # Check command patterns
        if self._is_normal_command_pattern(context.get("recent_commands", [])):
            score += 10

        # Check timing patterns
        if self._is_normal_timing_pattern(context.get("command_timestamps", [])):
            score += 10

        # Check resource usage
        if self._is_normal_resource_usage(context.get("resource_metrics", {})):
            score += 10

        return score

    async def _validate_context(self, context: Dict) -> int:
        """Validate request context"""
        score = 0

        # Network context
        if self._is_valid_network_context(context.get("network", {})):
            score += 10

        # Time context
        if self._is_valid_time_context(context.get("time", {})):
            score += 10

        # Location context
        if self._is_valid_location_context(context.get("location", {})):
            score += 10

        return score

    def generate_session_token(
        self, bot_id: str, trust_score: int, context: Dict
    ) -> str:
        """Generate a short-lived session token"""
        payload = {
            "bot_id": bot_id,
            "trust_score": trust_score,
            "context_hash": self._hash_context(context),
            "exp": time.time() + 3600,  # 1 hour expiration
        }

        return jwt.encode(payload, self.jwt_secret, algorithm="HS256")

    def verify_session_token(
        self, token: str, context: Dict
    ) -> Tuple[bool, Optional[str]]:
        """Verify session token and context"""
        try:
            payload = jwt.decode(token, self.jwt_secret, algorithms=["HS256"])

            # Verify context hasn't changed significantly
            current_context_hash = self._hash_context(context)
            if payload["context_hash"] != current_context_hash:
                return False, "Context mismatch"

            # Verify trust score hasn't decreased
            current_trust_score = self.trust_scores.get(payload["bot_id"], 0)
            if current_trust_score < payload["trust_score"] * 0.8:
                return False, "Trust score decreased"

            return True, None

        except jwt.ExpiredSignatureError:
            return False, "Token expired"
        except jwt.InvalidTokenError:
            return False, "Invalid token"

    def _hash_context(self, context: Dict) -> str:
        """Create a hash of context for comparison"""
        context_str = json.dumps(context, sort_keys=True)
        return hashlib.sha256(context_str.encode()).hexdigest()

    def _is_normal_command_pattern(self, commands: List[str]) -> bool:
        """Check if command sequence follows normal patterns"""
        if not commands:
            return True
        # Add pattern recognition logic here (e.g., detect loops or dangerous sequences)
        return True

    def _is_normal_timing_pattern(self, timestamps: List[float]) -> bool:
        """Check if command timing follows normal patterns"""
        if not timestamps:
            return True
        # Add timing analysis logic here (e.g., detect bot flooding)
        return True

    def _is_normal_resource_usage(self, metrics: Dict) -> bool:
        """Check if resource usage is within normal ranges"""
        if not metrics:
            return True

        cpu_usage = metrics.get("cpu_usage", 0)
        memory_usage = metrics.get("memory_usage", 0)

        return cpu_usage < 80 and memory_usage < 80

    def _is_valid_network_context(self, network: Dict) -> bool:
        """Validate network context"""
        if not network:
            return False

        required_fields = {"ip", "protocol", "port"}
        return all(field in network for field in required_fields)

    def _is_valid_time_context(self, time_context: Dict) -> bool:
        """Validate time-based context"""
        if not time_context:
            return False

        current_hour = datetime.datetime.now().hour
        allowed_hours = time_context.get("allowed_hours", range(24))

        return current_hour in allowed_hours

    def _is_valid_location_context(self, location: Dict) -> bool:
        """Validate location context"""
        if not location:
            return False

        required_fields = {"country", "region"}
        return all(field in location for field in required_fields)
