DEV/orchestrator/security/requirements-security.txt
# FilePath: "/DEV/orchestrator/security/requirements-security.txt"
# Project: Unified Bot Protocol (UBP) - Reference Implementation
# Component: Security (Zero Trust, encryption, authN/Z)
# Author: "Michael Landbo"
# Date created: "21/12/2025"
# Date Modified: "21/12/2025"
# Version: "v.1.0.0"
#
# License: Apache-2.0
#
# Notes:
# - Shared security libs for encryption, signing, and structured security logging.

cryptography
PyJWT
structlog
